+++ apps/backoffice-v2/CHANGELOG.md
@@ -8,6 +8,7 @@
   - @ballerine/common@0.9.6
   - @ballerine/workflow-browser-sdk@0.6.9
   - @ballerine/workflow-node-sdk@0.6.9
+  - @ballerine/blocks@0.2.3
 
 ## 0.7.7